* Sinatra bugfix: `mustache :"TEMPLATE.atom"`
Previously this would work in unexpected ways. Now it
behaves thusly:

Given `mustache :"positions.atom"`, Mustache will try
to find `Views::Positions::Atom`. If it finds it, it'll
use it. If it can't find it, it'll create it as a subclass
of `Views::Positions` and set the `template_name` setting to
"positions.atom".