Fixed an error in the docs regarding empty lists.
The documentation states that if a key is present and it
has a value of false or an empty list, then the content
between the pound sign and slash will not be shown.

The documentation had the key "person" with a true value,
which means that what {{#person}} {{/person}} was wrapping
would infact be shown.  The output however did not include
the contents of what {{#person}} {{/person}} wrapped.

If person is true, then the content should be shown, if
false, then not.

Changed the person key in the hash to be false and the
template to check if 'person' is empty/false to provide
a clear example.

Changed the template example for false/empty lists

Changed the template to check if 'person' is empty rather
than 'nothin' which was never defined. This shows a more
accurate and clear example of how a false or empty value
will result in the section not being shown.

Removed extra comma from hash.