Benchmark for performance of partial rendering
When mustache encounters `{{> partial_name}}` it loads the
`partial_name` partial and renders it using the current context.
This is a fairly expensive code path because it involves loading the
partial template file, compiling the template, and rendering it.

In order to make improvements to the performance of this we want to have
a benchmark to assess code changes against.