Add magic frozen_string_literal: false comments to source files
In order to make the code not emit warnings with ruby 3.4.x and
compatible with 3.5.x magic frozen_string_literal: false comments are
added to the source files.

In the future the code could be audited and changed to be compatible
with frozen string literals. A known incompatible line is in
lib/generator.rb:

```ruby
        exp[1..-1].reduce("") { |sum, e| sum << compile!(e) }
```

See: https://bugs.ruby-lang.org/issues/20205

> Workflow for library maintainers
> As a library maintainer, fixing the deprecation warnings can be as simple as prepending # frozen_string_literal: false at the top of all their source files, and this will keep working forever.
>
> Alternatively they can of course make their code compatible with frozen string literals.
>
> Code that is frozen string literal compatible doesn't need to explicitly declare it. Only code that need it turned of need to do so.